Move `Query`
