+++ crates/preset_env_base/Cargo.toml
@@ -12,6 +12,7 @@ version = "0.1.0"
 [dependencies]
 anyhow = "1"
 browserslist-rs = "=0.8.0"
+from_variant = {version = "0.1.3", path = "../from_variant"}
 semver = {version = "1.0.4", features = ["serde"]}
 serde = {version = "1", features = ["derive"]}
 st-map = "0.1.2"
